poll/select: Use 64bit serial for poll/select's kevent.udata.
This fixes the issue mentioned in this commit:
ce4975442fa0524017fb3c1aef93bbe6880ae770

It takes ~200 years for 2.5Ghz cpu to make the 64bit serial wrap,
even if the cpu's speed were 10 times faster tomorrow, it still
would take two decades to make the 64bit serial wrap.

Suggested-by: dillon@

+++ sys/sys/proc.h
@@ -225,7 +225,7 @@ struct lwp {
 	struct thread	*lwp_thread;	/* backpointer to proc's thread */
 	void		*lwp_unused01;	/* for future fields */
 	struct kqueue	lwp_kqueue;	/* for select/poll */
-	u_int		lwp_kqueue_serial;
+	uint64_t	lwp_kqueue_serial; /* for select/poll */
 	struct lwkt_token lwp_token;	/* per-lwp token for signal/state */
 	struct spinlock lwp_spin;	/* spinlock for signal handling */
 	void		*lwp_reserveds1; /* reserved for lwp_saveusp */
